feat: add menu
