+++ src/components/overlays/AuthValidator.tsx
@@ -25,10 +25,10 @@ export default function AuthValidator() {
 
   if (userInfoQuery.status !== 'success' && router.pathname !== '/login')
     return (
-      <div className='fixed inset-0 z-50 grid place-items-center bg-amber-400'>
-        <div className='flex items-center gap-2 text-lg tracking-widest'>
+      <div className='fixed inset-0 z-50 grid place-items-center bg-teal-700'>
+        <div className='flex items-center gap-2 text-gray-900'>
           <Spinner className='h-5 w-5' />
-          身分驗證中
+          <p className='text-lg font-bold tracking-widest'>身分驗證中</p>
         </div>
       </div>
     )
