+++ src/pages/_app.tsx
@@ -2,11 +2,11 @@ import type { AppType } from 'next/app'
 
 import trpc from '@/lib/client/trpc'
 import { useStore } from '@/lib/client/store'
-import EventListener from '@/components/EventListener'
-import AuthValidator from '@/components/AuthValidator'
-import Notification from '@/components/Notification'
-import Menu from '@/components/Menu'
-import RouterProgress from '@/components/RouteProgress'
+import EventListener from '@/components/overlays/EventListener'
+import AuthValidator from '@/components/overlays/AuthValidator'
+import Notification from '@/components/overlays/Notification'
+import Menu from '@/components/overlays/Menu'
+import RouterProgress from '@/components/overlays/RouteProgress'
 import '@/styles/globals.css'
 
 const PageApp: AppType = ({ Component, pageProps }) => {
@@ -14,12 +14,12 @@ const PageApp: AppType = ({ Component, pageProps }) => {
 
   return (
     <>
+      {/* Menu */}
+      {user && <Menu />}
       {/* Main */}
       <div className='mx-auto min-h-full max-w-lg'>
         <Component {...pageProps} />
       </div>
-      {/* Menu */}
-      <Menu />
       {/* Overlay */}
       <RouterProgress />
       {user && <EventListener />}
