+++ src/pages/orders.tsx
@@ -1,0 +1,7 @@
+export default function PageOrders() {
+  return (
+    <>
+      <h1 className='text-center text-2xl'>Orders</h1>
+    </>
+  )
+}
