+++ src/pages/reserve.tsx
@@ -1,0 +1,7 @@
+export default function PageReserve() {
+  return (
+    <>
+      <h1 className='text-center text-2xl'>Reserve</h1>
+    </>
+  )
+}
