+++ src/pages/wallet.tsx
@@ -1,0 +1,7 @@
+export default function PageWallet() {
+  return (
+    <>
+      <h1 className='text-center text-2xl'>Wallet</h1>
+    </>
+  )
+}
