+++ src/styles/globals.css
@@ -3,7 +3,7 @@
 @tailwind utilities;
 
 body {
-  @apply bg-stone-100 lg:overflow-y-scroll;
+  @apply bg-gray-200 text-gray-900;
 }
 
 html {
@@ -11,9 +11,9 @@ html {
 }
 
 #nprogress .bar {
-  @apply h-[3px] bg-stone-800;
-  box-shadow: 0 0 8px theme('colors.stone.800/50'),
-    0 0 5px theme('colors.stone.800');
+  @apply h-[3px] bg-teal-500;
+  box-shadow: 0 0 8px theme('colors.transparent'),
+    0 0 5px theme('colors.transparent');
 }
 #nprogress .peg {
   display: none;
